+++ web/src/features/fbaCalculator/components/FBATable.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { difference, filter, findIndex, isEmpty, isUndefined } from 'lodash'
 import {
-  Checkbox,
   FormControl,
   LinearProgress,
   makeStyles,
@@ -10,9 +9,7 @@ import {
   TableBody,
   TableCell,
   TableContainer,
-  TableHead,
-  TableRow,
-  TableSortLabel
+  TableRow
 } from '@material-ui/core'
 import GetAppIcon from '@material-ui/icons/GetApp'
 import { CsvBuilder } from 'filefy'
@@ -49,6 +46,7 @@ import ErrorAlert from 'features/fbaCalculator/components/ErrorAlert'
 import LoadingIndicatorCell from 'features/fbaCalculator/components/LoadingIndicatorCell'
 import SelectionCell from 'features/fbaCalculator/components/SelectionCell'
 import StickyCell from 'features/fbaCalculator/components/StickyCell'
+import FBATableHead from 'features/fbaCalculator/components/FBATableHead'
 
 export interface FBAInputGridProps {
   testId?: string
@@ -101,9 +99,6 @@ const useStyles = makeStyles(theme => ({
   grassCure: {
     width: 80
   },
-  windSpeed: {
-    width: 80
-  },
   paper: {
     width: '100%'
   },
@@ -120,8 +115,10 @@ const useStyles = makeStyles(theme => ({
     paddingLeft: '8px',
     paddingRight: '8px'
   },
-  tableHeaderRow: {
-    padding: '8px'
+  stickyProgress: {
+    left: 0,
+    position: 'sticky',
+    zIndex: theme.zIndex.appBar + 2
   }
 }))
 
@@ -402,319 +399,16 @@ const FBATable = (props: FBAInputGridProps) => {
         <div className={classes.display} data-testid={props.testId}>
           <Paper className={classes.paper} elevation={1}>
             <TableContainer className={classes.tableContainer}>
-              {loading && <LinearProgress />}
+              {loading && <LinearProgress className={classes.stickyProgress} />}
               <Table size="small" stickyHeader aria-label="Fire Behaviour Analysis table">
-                <TableHead>
-                  <TableRow>
-                    <StickyCell left={0} zIndexOffset={2}>
-                      <Checkbox
-                        data-testid="select-all"
-                        color="primary"
-                        checked={headerSelected}
-                        onClick={() => {
-                          if (headerSelected) {
-                            // Toggle off
-                            setSelected([])
-                            setHeaderSelect(false)
-                          } else {
-                            setSelected(
-                              rows.filter(row => !isUndefined(row)).map(row => row.id)
-                            )
-                            setHeaderSelect(true)
-                          }
-                        }}
-                      />
-                    </StickyCell>
-                    <TableCell key="header-zone" sortDirection={order}>
-                      <TableSortLabel
-                        className={classes.tableHeaderRow}
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Zone)
-                        }}
-                      >
-                        Zone
-                      </TableSortLabel>
-                    </TableCell>
-                    <StickyCell left={50} zIndexOffset={2}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Station)
-                        }}
-                      >
-                        Weather Station
-                      </TableSortLabel>
-                    </StickyCell>
-                    <TableCell key="header-elevation" sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Elevation)
-                        }}
-                      >
-                        Elev.
-                        <br />
-                        (m)
-                      </TableSortLabel>
-                    </TableCell>
-                    <StickyCell left={280} zIndexOffset={2}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => toggleSorting(SortByColumn.FuelType)}
-                      >
-                        FBP Fuel Type
-                      </TableSortLabel>
-                    </StickyCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => toggleSorting(SortByColumn.GrassCure)}
-                      >
-                        Grass
-                        <br />
-                        Cure
-                        <br />
-                        (%)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Status)
-                        }}
-                      >
-                        Status
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Temperature)
-                        }}
-                      >
-                        Temp
-                        <br />
-                        (&deg;C)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.RelativeHumidity)
-                        }}
-                      >
-                        RH
-                        <br />
-                        (%)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.WindDirection)
-                        }}
-                      >
-                        Wind
-                        <br />
-                        Dir
-                        <br />
-                        (&deg;)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell className={classes.windSpeed} sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.WindSpeed)
-                        }}
-                      >
-                        Wind Speed (km/h)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.Precipitation)
-                        }}
-                      >
-                        Precip
-                        <br />
-                        (mm)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.FFMC)
-                        }}
-                      >
-                        FFMC
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.DMC)
-                        }}
-                      >
-                        DMC
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.DMC)
-                        }}
-                      >
-                        DC
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.ISI)
-                        }}
-                      >
-                        ISI
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.BUI)
-                        }}
-                      >
-                        BUI
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.BUI)
-                        }}
-                      >
-                        FWI
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.HFI)
-                        }}
-                      >
-                        HFI
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.CriticalHours4000)
-                        }}
-                      >
-                        Critical
-                        <br />
-                        Hours
-                        <br />
-                        (4000 kW/m)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.CriticalHours10000)
-                        }}
-                      >
-                        Critical
-                        <br />
-                        Hours
-                        <br />
-                        (10000 kW/m)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.ROS)
-                        }}
-                      >
-                        ROS
-                        <br />
-                        (m/min)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.FireType)
-                        }}
-                      >
-                        Fire Type
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.CFB)
-                        }}
-                      >
-                        CFB (%)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.FlameLength)
-                        }}
-                      >
-                        Flame <br />
-                        Length <br /> (m)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.ThirtyMinFireSize)
-                        }}
-                      >
-                        30 min <br />
-                        fire size <br />
-                        (hectares)
-                      </TableSortLabel>
-                    </TableCell>
-                    <TableCell sortDirection={order}>
-                      <TableSortLabel
-                        direction={order}
-                        onClick={() => {
-                          toggleSorting(SortByColumn.SixtyMinFireSize)
-                        }}
-                      >
-                        60 min <br />
-                        fire size <br />
-                        (hectares)
-                      </TableSortLabel>
-                    </TableCell>
-                  </TableRow>
-                </TableHead>
+                <FBATableHead
+                  toggleSorting={toggleSorting}
+                  order={order}
+                  rows={rows}
+                  headerSelected={headerSelected}
+                  setHeaderSelect={setHeaderSelect}
+                  setSelected={setSelected}
+                />
                 <TableBody data-testid="fba-table-body">
                   {rows.map(row => {
                     return (
